Module route_guide_server::routeguide

Modules

client

server

Structs

Feature
A feature names something at a given point.

Point
Points are represented as latitude-longitude pairs in the E7 representation (degrees multiplied by 10**7 and rounded to the nearest integer). Latitudes should be in the range +/- 90 degrees and longitude should be in the range +/- 180 degrees (inclusive).

Rectangle
A latitude-longitude rectangle, represented as two diagonally opposite points "lo" and "hi".

RouteNote
A RouteNote is a message sent while at a given point.

RouteSummary
A RouteSummary is received in response to a RecordRoute rpc.