Trait route_guide_server::routeguide::server::RouteGuide

pub trait RouteGuide: Clone {
    type GetFeatureFuture: Future<Item = Response<Feature>, Error = Status>;
    type ListFeaturesStream: Stream<Item = Feature, Error = Status>;
    type ListFeaturesFuture: Future<Item = Response<Self::ListFeaturesStream>, Error = Status>;
    type RecordRouteFuture: Future<Item = Response<RouteSummary>, Error = Status>;
    type RouteChatStream: Stream<Item = RouteNote, Error = Status>;
    type RouteChatFuture: Future<Item = Response<Self::RouteChatStream>, Error = Status>;
    fn get_feature(&mut self, request: Request<Point>) -> Self::GetFeatureFuture;

    fn list_features(
        &mut self, 
        request: Request<Rectangle>
    ) -> Self::ListFeaturesFuture;

    fn record_route(
        &mut self, 
        request: Request<Streaming<Point>>
    ) -> Self::RecordRouteFuture;

    fn route_chat(
        &mut self, 
        request: Request<Streaming<RouteNote>>
    ) -> Self::RouteChatFuture;
}

Interface exported by the server.

Associated Types

type GetFeatureFuture: Future<Item = Response<Feature>, Error = Status>

type ListFeaturesStream: Stream<Item = Feature, Error = Status>

type ListFeaturesFuture: Future<Item = Response<Self::ListFeaturesStream>, Error = Status>

type RecordRouteFuture: Future<Item = Response<RouteSummary>, Error = Status>

type RouteChatStream: Stream<Item = RouteNote, Error = Status>

type RouteChatFuture: Future<Item = Response<Self::RouteChatStream>, Error = Status>

Loading content...
Required methods

fn get_feature(&mut self, request: Request<Point>) -> Self::GetFeatureFuture

A simple RPC.

Obtains the feature at a given position.

A feature with an empty name is returned if there's no feature at the given position.

fn list_features(
&mut self,
request: Request<Rectangle>
) -> Self::ListFeaturesFuture

A server-to-client streaming RPC.

Obtains the Features available within the given Rectangle. Results are streamed rather than returned at once (e.g. in a response message with a repeated field), as the rectangle may cover a large area and contain a huge number of features.

fn record_route(
&mut self,
request: Request<Streaming<Point>>
) -> Self::RecordRouteFuture

A client-to-server streaming RPC.

Accepts a stream of Points on a route being traversed, returning a RouteSummary when traversal is completed.

fn route_chat(
&mut self,
request: Request<Streaming<RouteNote>>
) -> Self::RouteChatFuture

A Bidirectional streaming RPC.

Accepts a stream of RouteNotes sent while a route is being traversed, while receiving other RouteNotes (e.g. from other users).

Loading content...